pyproject.toml
[project]
name = "hexlet-vacancy-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "django-filter>=25.1",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.10",
    "telethon>=1.40.0",
]

[dependency-groups]
dev = [
    "flake8>=7.2.0",
]
